Factored out Suggester class for cleaner design. - Eliminates deprecation warnings about getServiceLocator().
