+++ module/VuFind/src/VuFind/Autocomplete/PluginManager.php
@@ -48,67 +48,4 @@ class PluginManager extends \VuFind\ServiceManager\AbstractPluginManager
     {
         return 'VuFind\Autocomplete\AutocompleteInterface';
     }
-
-    /**
-     * This returns an array of suggestions based on current request parameters.
-     * This logic is present in the factory class so that it can be easily shared
-     * by multiple AJAX handlers.
-     *
-     * @param \Zend\Stdlib\Parameters $request    The user request
-     * @param string                  $typeParam  Request parameter containing search
-     * type
-     * @param string                  $queryParam Request parameter containing query
-     * string
-     *
-     * @return array
-     */
-    public function getSuggestions($request, $typeParam = 'type', $queryParam = 'q')
-    {
-        // Process incoming parameters:
-        $type = $request->get($typeParam, '');
-        $query = $request->get($queryParam, '');
-        $searcher = $request->get('searcher', 'Solr');
-        $hiddenFilters = $request->get('hiddenFilters', []);
-
-        // If we're using a combined search box, we need to override the searcher
-        // and type settings.
-        if (substr($type, 0, 7) == 'VuFind:') {
-            list(, $tmp) = explode(':', $type, 2);
-            list($searcher, $type) = explode('|', $tmp, 2);
-        }
-
-        // get Autocomplete_Type config
-        $options = $this->getServiceLocator()
-            ->get('VuFind\SearchOptionsPluginManager')->get($searcher);
-        $config = $this->getServiceLocator()->get('VuFind\Config')
-            ->get($options->getSearchIni());
-        $types = isset($config->Autocomplete_Types) ?
-            $config->Autocomplete_Types->toArray() : [];
-
-        // Figure out which handler to use:
-        if (!empty($type) && isset($types[$type])) {
-            $module = $types[$type];
-        } elseif (isset($config->Autocomplete->default_handler)) {
-            $module = $config->Autocomplete->default_handler;
-        } else {
-            $module = false;
-        }
-
-        // Get suggestions:
-        if ($module) {
-            if (strpos($module, ':') === false) {
-                $module .= ':'; // force colon to avoid warning in explode below
-            }
-            list($name, $params) = explode(':', $module, 2);
-            $handler = $this->get($name);
-            $handler->setConfig($params);
-        }
-
-        if (is_callable([$handler, 'addFilters'])) {
-            $handler->addFilters($hiddenFilters);
-        }
-
-        return (isset($handler) && is_object($handler))
-            ? array_values($handler->getSuggestions($query)) : [];
-    }
 }
